Added a setting field to the question object to choose between interviews and panels
Added a new custom field to question called setting, refactored the question creation in the interview admin tab to include this field and create the object with its controller There is a piece of validation that wont allow soft skills to be saved as a panel but maybe this could be better as like a dependent picklist scenario and got rid of the extra question creation tab. Updated the question pool list to include the setting. Also created a separate controller for the mock panel creation because they had one controller for interviews and panels(basically the only difference rn is a AND clause in 2 select statements so maybe this could be put back into one but this was the easiest way).